chore: moved merge-config file position
