+++ apps/web/src/lib/config/global-config/helpers/merge-config.ts
